Added more folder structure
