+++ backend/conduit/organizations/models.py
@@ -1,0 +1,8 @@
+import datetime as dt
+
+from flask_jwt_extended import current_user
+from slugify import slugify
+
+from conduit.database import (Model, SurrogatePK, db, Column,
+                              reference_col, relationship)
+from conduit.profile.models import UserProfile
