+++ backend/conduit/tags/models.py
@@ -1,0 +1,22 @@
+# coding: utf-8
+
+import datetime as dt
+
+from flask_jwt_extended import current_user
+from slugify import slugify
+
+from conduit.database import (Model, SurrogatePK, db, Column,
+                              reference_col, relationship)
+
+
+class Tags(Model):
+    __tablename__ = 'tags'
+
+    id = db.Column(db.Integer, primary_key=True)
+    tagname = db.Column(db.String(100))
+
+    def __init__(self, tagname):
+        db.Model.__init__(self, tagname=tagname)
+
+    def __repr__(self):
+        return self.tagname
